Resolve imported resource paths
Until now only the name of an imported resource was considered when
resolving that resource. This leads to multiple resources identified for
the same import, especially for commonly named files like utils.robot,
setup.robot, ...

Path of an imported resource will be used along with its name to better
handle this situation.